Add dataflow CFG support for the result expr of a lambda.
This adds a new node to the cfg: LambdaResultExpressionNode,
which corresponds to the result of a lambda expression with a
single expression as its body. This node is added at the end
of the cfg when CFGBuilder is passed a CFGLambda of this kind
as the underlaying ast.

This is needed by some analyses to handle the result of a lambda
expression in the case where the body is a single ExpressionTree.
The case where the body is a block is already handled by matching
on ReturnNode.